feat(nn): add BCEWithLogits loss and improve module state handling
refactor(attention): optimize multi-head attention implementation
fix(streaming): remove unused sender fields in dataset streams
perf(autograd): implement proper NLL loss backward pass
style(nn): clean up unused imports and fix warnings
docs(pycoeus): update Python API exports and fix circular imports
test(nn): add state_dict verification tests
build: fix build errors and warnings

+++ nn/src/rms_norm.rs
@@ -261,6 +261,7 @@ mod tests {
     use backend::CpuBackend;
     use dtype::float::Float32;
     use storage::DenseStorage;
+    use num_traits::Float;
 
     type TestBackend = CpuBackend<Float32>;
     type TestStorage = DenseStorage<Float32>;
@@ -356,8 +357,8 @@ mod tests {
         let output_data = output.as_slice();
 
         // Should be same as with affine but weight=1
-        let expected_rms = ((1.0 + 4.0 + 9.0 + 16.0) / 4.0).sqrt();
-        assert!((output_data[0] - Float32::new(1.0) / Float32::new(expected_rms as f32)).abs() < Float32::new(1e-5));
+        let expected_rms = ((1.0_f32 + 4.0 + 9.0 + 16.0) / 4.0).sqrt();
+        assert!((output_data[0] - Float32::new(1.0) / Float32::new(expected_rms)).abs() < Float32::new(1e-5));
     }
 
     #[test]
